+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Removed
 - silo-oracles: remove heartbeat check from DIA oracle
+- silo-oracles: remove heartbeat check from Chainlink oracle
 
 ### [3.7.0] - 2025-06-10
 ### Added